Applying MVC Design 01
Issue of current project file :
ViewController is big and long.
(Which means it has too many functions and variables.)
-> in need of refactoring

Hence,
Applying MVC Pattern to the project
Create Models, Views, Controllers groups
And create a new file BMICalculatorManager in Models
Some logics which were in ViewController, now in BMICalculatorManager to handle the business.